+++ src/main.js
@@ -5,11 +5,17 @@ import ts from 'typescript'
 import yargs from 'yargs'
 import { hideBin } from 'yargs/helpers'
 // internal dependencies
-import GetPathCommand from './command/get-path-command'
+import GetPathCommand from './get-path/get-path-command'
 import TypescriptHelper from './helpers/typescript-helper'
+import { DiffCollectionFactory } from './map-diffs-to-paths/diff-collection'
+import MapDiffsToPathsCommand from './map-diffs-to-paths/map-diffs-to-paths-command'
 
 const tsHelper = new TypescriptHelper(path, ts)
+// set up the getPath command
 const getPathCommandInstance = new GetPathCommand(fs, tsHelper)
+// set up the mapDiffsToPaths command
+const diffCollectionFactory = new DiffCollectionFactory()
+const mapDiffsToPathsCommandInstance = new MapDiffsToPathsCommand(fs, tsHelper, diffCollectionFactory)
 
 export default yargs(hideBin(process.argv))
 	.usage('Usage: $0 <command> [options]')
@@ -19,7 +25,13 @@ export default yargs(hideBin(process.argv))
 		builder: GetPathCommand.builder,
 		handler: getPathCommandInstance.handler.bind(getPathCommandInstance),
 	})
-	.demandCommand(1, 1, `choose a command: ${GetPathCommand.command}`)
+	.command({
+		command: MapDiffsToPathsCommand.command,
+		describe: MapDiffsToPathsCommand.describe,
+		builder: MapDiffsToPathsCommand.builder,
+		handler: mapDiffsToPathsCommandInstance.handler.bind(mapDiffsToPathsCommandInstance),
+	})
+	.demandCommand(1, 1, `choose a command: ${GetPathCommand.command} or ${MapDiffsToPathsCommand.command}`)
 	.strict()
 	.version('v')
 	.alias('v', 'version')