+++ CHANGELOG.md
@@ -8,4 +8,5 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ### Added
 
-- Add `get-path` command to get the path of a symbol in a file at a specific line.
+- Add `get-path` command to get the path to the symbol in a file at a specific line.
+- Add `map-diffs-to-paths` command to maps diffs to their corresponding symbol paths.